Introduce Strict vs Lazy code. Separate pattern matching stuff in the `duality` module.
