+++ src/duality/mod.rs
@@ -1,0 +1,4 @@
+pub mod value;
+pub mod pattern;
+pub mod matcher;
+pub mod sender;
